hmm test not passing on linux

+++ src/gdgem/tests/HitGenerator.h
@@ -38,8 +38,8 @@ namespace Gem {
     /// \todo Should deadtime comparison be strictly less than or less than or equal?
     /// For now it is strictly less than. Hence (oldtime - time) == 1 is within DT
     /// and (OT - NT) == 2 is outside DT if DT == 2
-    std::vector<Hit> & makeHitsForSinglePlane(uint8_t Plane, uint8_t MaxHits,
-         uint16_t X0, uint16_t Y0, float Degrees, uint8_t Gaps, uint32_t DeadTime, bool Shuffle);
+    std::vector<Hit> & makeHitsForSinglePlane(uint8_t Plane, int MaxHits,
+         int X0, int Y0, float Degrees, uint8_t Gaps, uint32_t DeadTime, bool Shuffle);
 
     /// \brief debug and testing function
     void printHits();
@@ -57,8 +57,8 @@ namespace Gem {
     uint64_t T0{0}; // ns
     uint32_t TGap{100}; // ns between events
     uint32_t dT{1}; // ns between Hits
-    uint16_t CoordMin{0};
-    uint16_t CoordMax{1279};
+    int CoordMin{0};
+    int CoordMax{1279};
 
     std::vector<Hit> Hits;
     std::mt19937 RandGen; ///< mersenne twister random number generator
